Added tests to cover all branches in delta coverage
